requirements.txt
pyrogram>=2.0.0
tgcrypto>=1.2.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncpg>=0.29.0
aioboto3>=12.3.0